isInheritorOfCollection logic moved into TypeResolver
